Implement Revoking a RefreshToken
This is a simple way to wipeOut all the previous sessions created by the user. Note that all the prevous accesstoken that the inital refreshToken created before it was wipped out will still work but it will be for a limited time

SIDENOTE: I had to refactor my db a bit. So it carried over to some files

AFTER this commit,
- We have an enpoint, where the user can get an accessToken and refresh token
- We have an enpoint the user can invoke to get a new accessToken. Remember accesstoken lasts for a short while. eg 30mins. RefreshToken lasts for a longer period
- We have an a method that the user can call to revoke all previous refreshTokens. We did not expose this function to a API though