Create a WebSocket API with APIGateway
We implemented webSockets so that the user will not have to be making fetch requests over and over again to get new invoices uploaded by colleagues. We will send notifications from our app to our clients. Clients can send messages as well through this persistent connection as well

The idea here is we have a dynamoDb to store list of connectionIds. These Ids are users that are currently connected to our webSocket. We basically send data to all these connected ids via our webSocket endpoint from whatever API(runs any lambda function) we want.

We have two Lambda functions that receive events for webSockets. We named these two function connect and disconnect.

The connect function recieve events for when a client connects to our webSocket via the webSocket url. The $connect event is directed to this function from the Gateway. So the get the client connection ID and save it in the server

The disconnect function receives the event of when a client disconnect from our webSocket. the $disconnect is sent to this function. So we get the client connection id and delete it from the table. Because the user is no more connected

NOTE: with the 'Websocket' persistent URL we get from the API Gateway when we define our 'Connection' URLS, we can
POST (send message to a client),
GET (get latest connection status) or
DELETE (disconnect a client from our API)

You will be only give one websocket url by gateway for your application

We used wscat npm package to test out web socket url wss://<domain_name>