Sending messages from websockets
We want want a user update request(adding an invoice record) to be decoupled from when we send the notifications.
A user we only make the update then the notifications to all other users will happen eventially

With the help of DynamoDB Updates Stream we can do this

More better practises to scale websockets in serverless and different way to implement websocket
https://www.serverless.com/blog/api-gateway-websockets-example
https://tsh.io/blog/implementing-websocket-with-aws-lambda-and-api-gateway/
https://www.serverless.com/framework/docs/providers/aws/events/websocket/

Some cases where you want to send push notifications SNS will be handy. Check out
SNS infohttps://stackoverflow.com/questions/32465505/cant-create-a-sns-event-source-on-a-lambda-function-using-cloudformation
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/AWS_SNS.html (all info an SNS cloud formation)
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-join.html (join function)
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/intrinsic-function-reference-getatt.html (get function)