Implement keeping the userlogged in on page refresh
The problem is, when the user, refreshed the page, the login sessin is not kept. The get shoot back to the login page.

We solved that problem in this commit by basically fetching a new accessToken anytime the user refreshed the page

FYI: i refactored my refreshToken and Login Lambda to return the refreshToken in in the API response body. The client, sets its cookie. More on why here
https://stackoverflow.com/questions/58588976/axios-is-not-sending-cookies