Implement Refreshing AccessToken
The idea here was any action dispatched in the store will likely require a network action as well, so for any actions dispatched, i check for if the accessToken is still valid.

If it is not valid, i get a new accessToken for them before we dispatch their action

Right now, if there is an error fetching anew accessToken(due to no accessToken), i just log them out. So that they can login again

If my app was to work offline, i might handle this differenlty, based on the error code or someome other way like writing to indexDb then synching later

FYI: I had to test in the addInvoice page as well as reducing the token time to confirm this works

I refactored the App.tsx due to some warning on having an empty slice of store returned