Implement loading initial data for the app
Implementing this required adding new slice of store state for invoices and users and action and reducers for these

We only load the initial data if the user is authenticated

FYI: i had to refactor the accessToken to only allow string and not 'null'.
Another alternative was to check if accessToken was not. But i thought reactoring to just string fits this case

Artivles that helped me made decisions
https://stackoverflow.com/questions/62516972/fetching-data-with-redux-thunk-middle-warehttps://stackoverflow.com/questions/62234869/reactjs-axios-interceptor-response-request
https://blog.logrocket.com/data-fetching-in-redux-apps-a-100-correct-approach-4d26e21750fc/