Implement addInvoice Action
Added, the action and reducers for these.

FYI: I had to refactor the backEnd a bit so we submit the date in the front-endinstead of the backend having to add the date attribute

UseFul resourses for this commit

For some reason, the input box converst the numbers to string. I fix that here
https://stackoverflow.com/questions/53133760/why-does-input-in-react-redux-form-change-a-number-into-a-string

You could searalize your input in front-end to avoid someone injecting js from input box
https://www.npmjs.com/package/serialize-javascript

More on Redux Reducers updates
https://stackoverflow.com/questions/33140008/react-redux-store-updates-but-react-does-not