Storing Secrets with AWS SecretManager and KMS
NOTE: the way I wrote iamRoleStatements is not depreciated. The new way to write these is in the link below. You may not need to add 'serverless-iam-roles-per-function' plugin to write permissions specifically to a function
https://www.serverless.com/framework/docs/providers/aws/guide/iam/

Helped generated encription keys!
https://www.allkeysgenerator.com/Random/Security-Encryption-Key-Generator.aspx

NOTE: I had to fix an unhandled error thrown when there is no cookie in front-end for refreshtoken