Implement Authentication using AWS Amplify
The steps i took to set up amplify auth is in the link below
https://blog.logrocket.com/authentication-react-apps-aws-amplify-cognito/

Basically Amplify AWS created a new userPool and userpoolclient for me. So Amplify creates the resources and also has API that you will in your Front-end for Components. Really good stuff. You can customize it as long as you want
So the one i created with serverless will is not used anymore. I will remove it later

More on the Amplify Authenticator UI
https://docs.amplify.aws/ui/auth/authenticator/q/framework/react#recommended-usage

More Related links
https://medium.com/@coryschimmoeller/customizing-the-authentication-experience-of-amplifys-withauthenticator-e6f2089ff469
https://docs.amplify.aws/lib/auth/customui/q/platform/js#customize-withauthenticator-hoc
https://medium.com/@kavyababu/customizing-aws-amplify-signin-page-in-react-94d95983d0c
https://github.com/aws-amplify/docs/issues/2208
https://blog.kylegalbraith.com/2018/11/29/how-to-easily-customize-the-aws-amplify-authentication-ui/
https://stackoverflow.com/questions/59270735/how-to-retrieve-cognito-userpool-attributes-using-getuser-react-js
https://stackoverflow.com/questions/59270735/how-to-retrieve-cognito-userpool-attributes-using-getuser-react-js