Implement Importing existing Cognito to Amplify
In this commit, i implemented using the cognito userpool we provision with serverless and imported it into amplify UI for the front end to use

**FYI: We did not do much in the front-end other that removing the old auth and importing our own exisiting auth cognito that serverless provisioned

I had to create a userPoolClient for native mobile apps as recommended by amplify
Then create an identity pool that have the two clients. For the identity pool, i created it from the console and not serverless cli
https://aws.amazon.com/blogs/mobile/use-existing-cognito-resources-for-your-amplify-api-storage-and-more/

The i removed the old auth resources tha amplify created for me and imported my own that serverless did
https://docs.amplify.aws/cli/auth/import

All the apis, you can expose for admin to use
https://docs.amplify.aws/cli/auth/admin

I also attached a lambda function to be triggered at post confirmation after a user signs up.
https://www.serverless.com/framework/docs/providers/aws/events/cognito-user-pool/#using-existing-pools

More ideas on what the post confirmation lambda could do
https://aws.amazon.com/es/blogs/mobile/amplify-framework-adds-supports-for-aws-lambda-triggers-in-auth-and-storage-categories/
https://stackoverflow.com/questions/49984024/aws-cognito-amplify-have-new-user-sign-ups-be-automatically-add-to-a-user-grou
https://blog.bitsrc.io/building-otp-authentication-with-reactjs-and-aws-amplify-c5fd2e517fac

How add user to a group in front-end by an onclick event
https://aws.amazon.com/blogs/mobile/amplify-cli-enables-creating-amazon-cognito-user-pool-groups-configuring-fine-grained-permissions-on-groups-and-adding-user-management-capabilities-to-applications/

How to check if current authenticated user is admin in front-end
https://stackoverflow.com/questions/63801796/aws-cognito-checking-if-authenticated-user-is-in-group-or-is-admin-using-amplif

**How to provision identity pool with serverless. For your project, you created it from console
https://github.com/mavi888/sam-cognito-auth/blob/master/backend/template.yml

References for how to configure a userpool resource even more that the simple one you have now
https://forum.serverless.com/t/cognito-user-identity-pools-as-serverless-yml-resource-defs/2050/2
https://github.com/nkhine/elm-golang-spa/blob/master/serverless.yml#L290
https://forum.serverless.com/t/linking-lambda-function-to-an-exiting-user-pool/4688/2
https://stackoverflow.com/questions/58277011/serverless-framework-lambda-function-with-authorizer-cognito-user-pools-always
https://forum.serverless.com/t/access-to-cloud-formation-output-of-a-cognito-user-pool-in-authorizer/11005/2